test ok normi ko

+++ ft_hexalen.c
@@ -6,7 +6,7 @@
 /*   By: psoto-go <psoto-go@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/02 14:09:15 by psoto-go          #+#    #+#             */
-/*   Updated: 2021/11/03 11:03:42 by psoto-go         ###   ########.fr       */
+/*   Updated: 2021/11/03 15:09:09 by psoto-go         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,13 @@
 
 int	ft_hexalen(unsigned long int n)
 {
-	int co;
-	int count;
-	
+	unsigned long int	co;
+	int					count;
+
 	co = 0;
 	count = 1;
-	while (n >= 16){
+	while (n >= 16)
+	{
 		co = n / 16;
 		n = co;
 		count++;
